fix crashes and root refactoring
